# BE/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LingoPrep API",
    "description": "Intelligent IELTS & Aptis Speaking & Writing Practice & Assessment Platform powered by AI",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/submissions/writing": {
      "post": {
        "tags": [
          "Submissions & Assessments"
        ],
        "summary": "Submit Writing",
        "operationId": "submit_writing_api_submissions_writing_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmissionCreateWriting"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/submissions/speaking": {
      "post": {
        "tags": [
          "Submissions & Assessments"
        ],
        "summary": "Submit Speaking",
        "operationId": "submit_speaking_api_submissions_speaking_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_submit_speaking_api_submissions_speaking_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/submissions": {
      "get": {
        "tags": [
          "Submissions & Assessments"
        ],
        "summary": "Get My Submissions",
        "operationId": "get_my_submissions_api_submissions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SubmissionResponse"
                  },
                  "type": "array",
                  "title": "Response Get My Submissions Api Submissions Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/submissions/{submission_id}": {
      "get": {
        "tags": [
          "Submissions & Assessments"
        ],
        "summary": "Get Submission",
        "operationId": "get_submission_api_submissions__submission_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "submission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Submission Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/submissions/{submission_id}/request-review": {
      "post": {
        "tags": [
          "Submissions & Assessments"
        ],
        "summary": "Request Teacher Review",
        "operationId": "request_teacher_review_api_submissions__submission_id__request_review_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "submission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Submission Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssessmentResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "submission_id": {
            "type": "integer",
            "title": "Submission Id"
          },
          "overall_band": {
            "type": "number",
            "title": "Overall Band"
          },
          "overall_cefr": {
            "type": "string",
            "title": "Overall Cefr"
          },
          "fluency_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fluency Score"
          },
          "lexical_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lexical Score"
          },
          "grammar_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grammar Score"
          },
          "pronunciation_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pronunciation Score"
          },
          "task_response_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Response Score"
          },
          "coherence_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coherence Score"
          },
          "criteria_breakdown": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Criteria Breakdown"
          },
          "strengths": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Strengths",
            "default": []
          },
          "weaknesses": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Weaknesses",
            "default": []
          },
          "inline_feedback": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Inline Feedback",
            "default": []
          },
          "model_answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Answer"
          },
          "recommendations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recommendations",
            "default": []
          },
          "evaluated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Evaluated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "submission_id",
          "overall_band",
          "overall_cefr",
          "evaluated_at"
        ],
        "title": "AssessmentResponse"
      },
...
      "Body_submit_speaking_api_submissions_speaking_post": {
        "properties": {
          "question_id": {
            "type": "integer",
            "title": "Question Id"
          },
          "duration_seconds": {
            "type": "integer",
            "title": "Duration Seconds",
            "default": 60
          },
          "transcript": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transcript"
          },
          "audio_file": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio File"
          }
        },
        "type": "object",
        "required": [
          "question_id"
        ],
        "title": "Body_submit_speaking_api_submissions_speaking_post"
      },
...
      "ExamType": {
        "type": "string",
        "enum": [
          "IELTS",
          "APTIS"
        ],
        "title": "ExamType"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QuestionResponse": {
        "properties": {
          "exam_type": {
            "$ref": "#/components/schemas/ExamType"
          },
          "skill": {
            "$ref": "#/components/schemas/SkillType"
          },
          "part": {
            "type": "string",
            "title": "Part"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "topic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic",
            "default": "General"
          },
          "difficulty": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Difficulty",
            "default": "Medium"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "instructions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instructions"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "prep_time_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prep Time Seconds",
            "default": 60
          },
          "time_limit_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Limit Seconds",
            "default": 120
          },
          "min_words": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Words"
          },
          "max_words": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Words"
          },
          "sample_answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample Answer"
          },
          "rubric_criteria": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rubric Criteria"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags",
            "default": []
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active",
            "default": true
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "exam_type",
          "skill",
          "part",
          "title",
          "prompt",
          "id",
          "created_at"
        ],
        "title": "QuestionResponse"
      },
...
      "SkillType": {
        "type": "string",
        "enum": [
          "SPEAKING",
          "WRITING"
        ],
        "title": "SkillType"
      },
      "SubmissionCreateWriting": {
        "properties": {
          "question_id": {
            "type": "integer",
            "title": "Question Id"
          },
          "content_text": {
            "type": "string",
            "title": "Content Text"
          }
        },
        "type": "object",
        "required": [
          "question_id",
          "content_text"
        ],
        "title": "SubmissionCreateWriting"
      },
      "SubmissionResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "question_id": {
            "type": "integer",
            "title": "Question Id"
          },
          "submission_type": {
            "$ref": "#/components/schemas/SkillType"
          },
          "content_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Text"
          },
          "audio_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Path"
          },
          "duration_seconds": {
            "type": "integer",
            "title": "Duration Seconds"
          },
          "word_count": {
            "type": "integer",
            "title": "Word Count"
          },
          "status": {
            "$ref": "#/components/schemas/SubmissionStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "question": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QuestionResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "assessment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssessmentResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "teacher_review": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TeacherReviewResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "question_id",
          "submission_type",
          "duration_seconds",
          "word_count",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "SubmissionResponse"
      },
      "SubmissionStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "PROCESSING",
          "EVALUATED",
          "REVIEW_REQUESTED",
          "REVIEWED"
        ],
        "title": "SubmissionStatus"
      },
...
      "TeacherReviewResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "submission_id": {
            "type": "integer",
            "title": "Submission Id"
          },
          "teacher_id": {
            "type": "integer",
            "title": "Teacher Id"
          },
          "teacher_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Teacher Name"
          },
          "overall_band": {
            "type": "number",
            "title": "Overall Band"
          },
          "overall_cefr": {
            "type": "string",
            "title": "Overall Cefr"
          },
          "criteria_scores": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Criteria Scores"
          },
          "teacher_notes": {
            "type": "string",
            "title": "Teacher Notes"
          },
          "is_overridden": {
            "type": "boolean",
            "title": "Is Overridden"
          },
          "reviewed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Reviewed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "submission_id",
          "teacher_id",
          "overall_band",
          "overall_cefr",
          "teacher_notes",
          "is_overridden",
          "reviewed_at"
        ],
        "title": "TeacherReviewResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        }
      }
    }
  },
...
}